backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "apscheduler>=3.10.0",
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "celery>=5.6.3",
    "fastapi>=0.135.1",
    "fastapi-mail>=1.6.2",
    "httpx>=0.28.1",
    "logger>=1.4",
    "loguru>=0.7.3",
    "massive>=2.4.0",
    "pandas>=3.0.1",
    "pydantic-ai[google,openai]>=1.88.0",
    "pydantic-settings>=2.13.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.2",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.22",
    "redis>=7.3.0",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.41.0",
    "yfinance>=1.2.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]
